requirements.txt
streamlit>=1.20
transformers>=4.30
torch>=1.13
Pillow>=9.0
numpy
opencv-python
imagehash
langchain>=0.0.320
google-generativeai>=0.2.0
fpdf2
google-search-results